refacto Grid and Period
